# Use memory map to speed up snapshot untar