--- a/runtime/Cargo.toml
+++ b/runtime/Cargo.toml
@@ -62,6 +62,9 @@ zstd = "0.11.1"
 crate-type = ["lib"]
 name = "solana_runtime"
 
+[target.'cfg(unix)'.dependencies]
+mmarinus = "0.4.0"
+
 [dev-dependencies]
 assert_matches = "1.5.0"
 ed25519-dalek = "=1.0.1"