--- a/runtime/src/shared_buffer_reader.rs
+++ b/runtime/src/shared_buffer_reader.rs
@@ -22,9 +22,9 @@ use {
 
 // tunable parameters:
 // # bytes allocated and populated by reading ahead
-const TOTAL_BUFFER_BUDGET_DEFAULT: usize = 4_000_000_000;
+const TOTAL_BUFFER_BUDGET_DEFAULT: usize = 2_000_000_000;
 // data is read-ahead and saved in chunks of this many bytes
-const CHUNK_SIZE_DEFAULT: usize = 200_000_000;
+const CHUNK_SIZE_DEFAULT: usize = 50_000_000;
 
 type OneSharedBuffer = Arc<Vec<u8>>;