--- a/runtime/src/snapshot_utils.rs
+++ b/runtime/src/snapshot_utils.rs
@@ -781,7 +781,7 @@ pub struct BankFromArchiveTimings {
 }
 
 // From testing, 4 seems to be a sweet spot for ranges of 60M-360M accounts and 16-64 cores. This may need to be tuned later.
-const PARALLEL_UNTAR_READERS_DEFAULT: usize = 4;
+const PARALLEL_UNTAR_READERS_DEFAULT: usize = 8;
 
 /// Rebuild bank from snapshot archives.  Handles either just a full snapshot, or both a full
 /// snapshot and an incremental snapshot.
@@ -1050,7 +1050,7 @@ where
         parallel_divisions,
     )?;
     measure_untar.stop();
-    info!("haha: {}", measure_untar);
+    info!("{}", measure_untar);
 
     let unpacked_version_file = unpack_dir.path().join("version");
     let snapshot_version = snapshot_version_from_file(&unpacked_version_file)?;
@@ -1451,6 +1451,7 @@ fn unpack_snapshot_local<T: 'static + Read + std::marker::Send, F: Fn() -> T>(
             unpack_snapshot(&mut archive, ledger_dir, account_paths, parallel_selector)
         })
         .collect::<Vec<_>>();
+
     let mut unpacked_append_vec_map = UnpackedAppendVecMap::new();
     for h in all_unpacked_append_vec_map {
         unpacked_append_vec_map.extend(h?);
@@ -1505,12 +1506,9 @@ fn untar_snapshot_in<P: AsRef<Path>>(
             account_paths,
             parallel_divisions,
         )?,
-        ArchiveFormat::Tar => unpack_snapshot_local(
-            || BufReader::new(slice),
-            unpack_dir,
-            account_paths,
-            parallel_divisions,
-        )?,
+        ArchiveFormat::Tar => {
+            unpack_snapshot_local(|| slice, unpack_dir, account_paths, parallel_divisions)?
+        }
     };
 
     Ok(account_paths_map)