#### Problem

https://github.com/solana-labs/solana/issues/24798

Use memory map to speed up snapshot untar.

#### Summary of Changes

1. use mmap to read snapshot archive files.
2. retune the parallel factor and chunksize

with mmap
```
reading entire decompressed file took: 44799890 us, bytes: 76654958592, read_us: 23072928, waiting_for_buffer_us: 21715846, largest fetch: 100000000, error: Ok(0)
```

master
```
reading entire decompressed file took: 44960804 us, bytes: 76654958592, read_us: 29821976, waiting_for_buffer_us: 15128175, largest fetch: 100000000, error: Ok(0)
```

The comparison does show that read_time is reduced from 29s to 23 seconds. But the the total untar time don't differ much. The reason is because, although we are reading the data faster, we spend more time on waiting for available buffer, 21s vs 15s. 

Next, we are going to tune the chunk ratio. 

```
// tunable parameters:
// # bytes allocated and populated by reading ahead
const TOTAL_BUFFER_BUDGET_DEFAULT: usize = 2_000_000_000;
// data is read-ahead and saved in chunks of this many bytes
const CHUNK_SIZE_DEFAULT: usize = 100_000_000;
```

```
parallel_factor=8 buf_size=2G chunk=50M
reading entire decompressed file took: 32857423 us, bytes: 76654958592, read_us: 21869626, waiting_for_buffer_us: 10964636, largest fetch: 50000000, error: Ok(0)

parallel_factor=8 buf_size=2G chunk=25M
reading entire decompressed file took: 33978930 us, bytes: 76654958592, read_us: 26325740, waiting_for_buffer_us: 7609328, largest fetch: 25000000, error: Ok(0)
```

So far, best config is *parallel_factor=8 buf_size=2G chunk=50M*. 
untar time reduced from 44s to 32s. 





Fixes #
<!-- OPTIONAL: Feature Gate Issue: # -->
<!-- Don't forget to add the "feature-gate" label -->