@jeffwashington 2022-05-03
what machine spec did you use to get these metrics?