@HaoranYi 2022-05-03
> what machine spec did you use to get these metrics?

gce box: 48core 250G RAM