@HaoranYi 2022-05-03
> How big does the slice become? Is it as big as the tar file? Will we run into physical memory limits with 1B/10B account snapshots?
In this test run, it is about 76G. 
Yeah. As long as we have enough *virtual memory* then it will be fine. Otherwise, it will throw an error  and exit.
```
    let mmap = Map::load(&snapshot_tar, Private, perms::Read).unwrap_or_else(|e| {
        error!(
            "Failed to map the snapshot file: {} {}.\n
                        Please increase the virtual memory on the system.",
            snapshot_tar.as_ref().display(),
            e,
        );
        std::process::exit(1);
```
Do you have an estimate of how big the snapshot will be for 1B/10B accounts?