@HaoranYi 2022-05-03
Or Perhaps, we can check the file size and switch to old BufReader if the size is too large?