@jeffwashington 2022-05-03
> Do you have an estimate of how big the snapshot will be for 1B/10B accounts?

it is compressed, but I imagine it to be basically linear. so if 20G: 170M accounts, then 200G: 1.7B accounts, 2T: 17B accounts. Other things will begin breaking.
I sometimes test on 64G machines. Perhaps that ship has sailed.