@HaoranYi 2022-05-04
Yes, will try.

On Wed, May 4, 2022 at 3:41 PM Jeff Washington (jwash) <
***@***.***> wrote:

> I'm happy to have this change made. I would love the perf improvements. I
> load from snapshots all day long.
> I am ignorant of how the slice code and the memory mapped file stuff will
> work in a mem env with very little virtual memory and slices mapped into
> memory. Does it page the file since it is file backed already and that
> doesn't count as vm in the same sense? So, will this just work?
> Can you write an app to hold all but say 15G of memory on a machine and
> leave it running. then try ledger-tool verify and see what hapens when we
> try to untar a 20G snapshot?
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/solana-labs/solana/pull/24889#issuecomment-1117898937>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AABVSJGU4IBDALAH3VLCV2LVILOF5ANCNFSM5U4E4K2Q>
> .
> You are receiving this because you authored the thread.Message ID:
> ***@***.***>
>