@sakridge 2022-05-04
> I'm happy to have this change made. I would love the perf improvements. I load from snapshots all day long.
> I am ignorant of how the slice code and the memory mapped file stuff will work in a mem env with very little virtual memory and slices mapped into memory. Does it page the file since it is file backed already and that doesn't count as vm in the same sense? So, will this just work?
> Can you write an app to hold all but say 15G of memory on a machine and leave it running. then try ledger-tool verify and see what hapens when we try to untar a 20G snapshot?

You mean physical not virtual right? The virtual limit is usually very high like many TB in Linux and you can adjust higher if needed.