@jeffwashington 2022-05-04
Well this is in my ignorance. We configure our validators with very little vm (2G)? If we mmap the tar file, then map it into a slice, what happens when we have less physical memory than we need to both represent the mmapped slice AND whatever other dynamic memory the validator needs, like the accounts index. I would assume the os would just page in the mmapped areas that are accessed from the file that backs it. I do not know if this counts in the vm limit or not. And, is the slice 20G elements long? Hopefully this 'just works'.

Basically, can I load a 80G snapshot on a 64G machine with 2G of virtual memory configured?
I suggested using up ram in a separate app so that we can answer the question:
Can I load a 20G snapshot on a machine with 2G of virtual memory and all but ~15G of memory used by another application.

If you know this will work, then I have no concerns when a 2B account snapshot is loaded on a 128G machine.