@t-nelson 2022-05-04
it should work, for some definition of "work".  mmap basically "swaps" out to the backing file if there's not enough RAM available.  Whether this is going to have a negative effect on memory pressure for other components is the question.  That is, I'm not sure if the in-memory region is treated as RES or more like VFS cache WRT eviction