@jeffwashington 2022-05-04
> Or Perhaps, we can switch to old BufReader if the memory map fails?

I could imagine the mmap succeeds, but it uses almost all of the available mem. Then, we'd oom generating the index or something. I would like generating the index and untarring the append vecs to be happening concurrently - something @HaoranYi and I have been talking about. I'm just trying to think ahead and not lay a trap for ourselves we hit later.