@jeffwashington 2022-05-04
> it should work, for some definition of "work". mmap basically "swaps" out to the backing file if there's not enough RAM available. Whether this is going to have a negative effect on memory pressure for other components is the question. That is, I'm not sure if the in-memory region is treated as RES or more like VFS cache WRT eviction

This summarizes my ignorance and concerns ;-) I also don't know. I think it also might 'work'. But I don't know enough low level details of linux memory management and mmapped files to 'know' with confidence. So, I proposed an experiment that seemed pretty easy. Feel free to propose a different experiment!