@jeffwashington 2022-05-04
> virtual memory can be specified. Usually it is set larger than the physical ram. For example, with 256G RAM, you can set it to vm to 512G. Then you application can map a file that is less than 512G, it should be fine. When you access the data, if it is not in physical ram, os will help to page in data into memory. But it may have bad performance because of thrashing of the available physical ram.

I do not know the history of validators. But, it appears that we set virtual memory at 2G on our machines. I do not know the history of this or the impact if we were to change it.