@sakridge 2022-05-04
The 2G is the swap file that is somewhat different from virtual memory and not really used when mmap'ing a file. Linux will page out to the mapped file in this case and keep recently used pages in physical memory in the page cache. The page cache itself could be maybe be paged out to swap, but swap is not necessary to map a huge vm space and the size of the swap has no bearing on how much you can map. I believe Linux keeps the page cache within the physical memory of the machine.