@HaoranYi 2022-05-04
yeah. The vm is set in the following file. In my case, it is 700G. 

```
haoran_yi_solana_com@hyi-dev2:~/git/hy_solana$ cat /etc/sysctl.conf | grep vm.max_map_count
vm.max_map_count=700000
```